refactor(app): name the restart retry policy after the job, not the channel
MicRestartRetryPolicy decides how long to wait before retrying a capture restart
and when to stop trying. Nothing in it is about microphones: it takes a count and
returns a delay or a give-up.

The next commit gives the app-audio channel the same budget, and a type called
Mic… used from the app-audio coordinator would be the kind of name that makes
the next reader assume two separate schedules exist. Renamed before the change
that shares it, so that commit is about behaviour rather than about text.

Pure rename, plus a note on the type saying why both channels share it: they
face the same event, and a device that needs three seconds to re-enumerate
should not cost one channel its track while the other rides it out.
