+++ tools/audiotap/Sources/CaptureRestartRetryPolicy.swift
@@ -1,19 +1,24 @@
 import Foundation
 
-/// Action to take after a mic-engine restart attempt failed.
-public enum MicRestartRetryAction: Equatable {
+/// Action to take after a capture restart attempt failed.
+public enum CaptureRestartRetryAction: Equatable {
     /// Retry the restart after the given backoff delay (seconds).
     case retry(afterSeconds: Double)
     /// Stop retrying — the failure budget is exhausted.
     case giveUp
 }
 
-/// Pure decision logic for retrying a failed mic-engine restart (issue #379).
+/// Pure decision logic for retrying a failed capture restart (issue #379).
 /// A device change can briefly expose an invalid format and make the restart
 /// throw; retrying with backoff lets a transient settle instead of dropping
-/// the recording. Separated from MicCaptureHandler so the backoff schedule and
-/// give-up boundary are unit-testable without hardware.
-public enum MicRestartRetryPolicy {
+/// the recording. Kept as a value type so the backoff schedule and the give-up
+/// boundary are unit-testable without hardware.
+///
+/// Shared by both channels on purpose. They face the same event, a device that
+/// went away and is coming back, and a device that needs three seconds to
+/// re-enumerate should not cost one channel its track while the other rides it
+/// out. Two schedules would drift apart at the first tuning change.
+public enum CaptureRestartRetryPolicy {
     /// Maximum number of retries before giving up.
     public static let maxAttempts = 5
 
@@ -27,7 +32,7 @@ public enum MicRestartRetryPolicy {
     ///   failure, 1 after one retry, …).
     /// - Returns: `.retry` with an exponentially-backed-off delay while within
     ///   budget, otherwise `.giveUp`.
-    public static func decide(attemptsSoFar: Int) -> MicRestartRetryAction {
+    public static func decide(attemptsSoFar: Int) -> CaptureRestartRetryAction {
         guard attemptsSoFar < maxAttempts else { return .giveUp }
         let delay = min(baseBackoff * pow(2.0, Double(attemptsSoFar)), maxBackoff)
         return .retry(afterSeconds: delay)
